Fix maintainance of invariant for length of runs
Timsort maintains a list of sorted segments (called 'runs') in an list
called 'runLen'. To ensure that a small array suffices for this, timsort
imposes an invariant on the lengths of runs: each run must be shorter
than the preceding run in 'runs' and the sum of the lengths of two
consecutive runs must be lower than the length of the run preceding the
two runs. In effect, the (reverted) sequence of run lengths grows at
least as fast as the fibonacci sequence. The operation that adds a new
run to the list of runs (and merges runs, if necessary) makes sure that
the new list 'runLen' satisfies

    runLen [n-2] > runLen [n-1] + runLen [n]
    runLen [n-1] > runLen [n]

It has recently been discovered that this doesn't suffice to maintain
the invariant for all triples of consecutive runs:

http://envisage-project.eu/proving-android-java-and-python-sorting-algorithm-is-broken-and-how-to-fix-it/

This patch fixes this problem.

Correctness of the new implementation has been checked with QuickCheck:
https://gist.github.com/timjb/cca12b004a0c782ca622