Halved memory usage of merge sort
The documentation at the top of `Merge.hs` claims "The temporary buffer
is preallocated to 1/2 the size of the input array", and this is now
correct.

We also save a few comparisons in order to be able to use one to avoid
allocating a vector if it won't be used at all.

Ceiling (`(len + 1)/2`) is used instead of floor (`len/2`) in order to
be able to fit the larger half of odd-length vectors.