Fix bug in Data.Vector.Algorithms.Heaps.heapInsert
The previous code confused the index in the original vector and the index
shifted by the offset (the `l` in the code), and it would lead to an
infinite loop if the offset is non-zero. In order to fix this issue,
this commit makes sure that all the index variable in the loop is shifted.